Attempt optimization for RVV max index
Attempt optimization for `volk_gnsssdr_8i_index_max_16u_rvv`
by not using a vectorized compare and instead just finding the
maximum of the vector selection and checking if it is less than
the preivously found maximum. This saves saves at least one
intensive vector instruction when there is a new max element in
this part of the vector.