Fix by remembering how pointer addition works
Fix by remembering that C is smart enough to automatically multiply
the addition of a number to a pointer by `sizeof(ptr)` so that
the memory access stays aligned and the correct element is accessed.

Previously, was extraneously multiplying by 4, resulting in a segfault
(probably from going way out of memory).