Try copying phase before multiplying
Try copying phase into its own vector register group before
multiplying to not have to target a phase vector register
group with a masked function. If the compiler was not smart
enough to implicitly copy beforehand, this may have
caused serious issues as try to use a phase vector register
group in a computation that had been partially written over
before it was supposed to be.