Implement 16ic 16i rotator dot product, puppet, with RVV
In process, attempted experimental optimization.
Specifically, at point where would perform dot product, loaded the
accumulators in a vectorized way so that can simply rotate through
the accumulators instead of having to reload each one at a time.

However, ended up actually making the test slower. This
may, however, be a symptom of the relatively small test, so
it may be worth keeping an eye out in case use cases
turn out to mainly be for bigger vectors.

Signed-off-by: Marcus Alagar <mvala079@gmail.com>